pyproject.toml
[project]
name = "agent-engineering-framework"
version = "0.1.0"
description = "AI engineering governance framework"
authors = [
    { name = "Arkimetrix Analytics Ltd." }
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<4.0,!=3.14.1"

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0.3,<7.0.0",
    "graphviz>=0.21,<0.22",
    "networkx>=3.6",
    "typer>=0.12",
    "rich>=13"
]

[tool.poetry]
packages = [{ include = "agent_framework" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
agent = "agent_framework.cli:main"